Revert "Update to latest 1.10.3 stable version"

+++ airflow/contrib/executors/kubernetes_executor.py
@@ -17,11 +17,10 @@
 
 import base64
 import hashlib
-from queue import Empty
-
 import re
 import json
 import multiprocessing
+from queue import Queue
 from dateutil import parser
 from uuid import uuid4
 import kubernetes
@@ -33,7 +32,8 @@ from airflow.contrib.kubernetes.kube_client import get_kube_client
 from airflow.contrib.kubernetes.worker_configuration import WorkerConfiguration
 from airflow.executors.base_executor import BaseExecutor
 from airflow.executors import Executors
-from airflow.models import KubeResourceVersion, KubeWorkerIdentifier, TaskInstance
+from airflow.models import TaskInstance
+from airflow.models.kubernetes import KubeResourceVersion, KubeWorkerIdentifier
 from airflow.utils.state import State
 from airflow.utils.db import provide_session, create_session
 from airflow import configuration, settings
@@ -43,16 +43,15 @@ from airflow.utils.log.logging_mixin import LoggingMixin
 
 class KubernetesExecutorConfig:
     def __init__(self, image=None, image_pull_policy=None, request_memory=None,
-                 request_cpu=None, limit_memory=None, limit_cpu=None, limit_gpu=None,
+                 request_cpu=None, limit_memory=None, limit_cpu=None,
                  gcp_service_account_key=None, node_selectors=None, affinity=None,
-                 annotations=None, volumes=None, volume_mounts=None, tolerations=None, labels=None):
+                 annotations=None, volumes=None, volume_mounts=None, tolerations=None):
         self.image = image
         self.image_pull_policy = image_pull_policy
         self.request_memory = request_memory
         self.request_cpu = request_cpu
         self.limit_memory = limit_memory
         self.limit_cpu = limit_cpu
-        self.limit_gpu = limit_gpu
         self.gcp_service_account_key = gcp_service_account_key
         self.node_selectors = node_selectors
         self.affinity = affinity
@@ -60,18 +59,17 @@ class KubernetesExecutorConfig:
         self.volumes = volumes
         self.volume_mounts = volume_mounts
         self.tolerations = tolerations
-        self.labels = labels or {}
 
     def __repr__(self):
         return "{}(image={}, image_pull_policy={}, request_memory={}, request_cpu={}, " \
-               "limit_memory={}, limit_cpu={}, limit_gpu={}, gcp_service_account_key={}, " \
+               "limit_memory={}, limit_cpu={}, gcp_service_account_key={}, " \
                "node_selectors={}, affinity={}, annotations={}, volumes={}, " \
-               "volume_mounts={}, tolerations={}, labels={})" \
+               "volume_mounts={}, tolerations={})" \
             .format(KubernetesExecutorConfig.__name__, self.image, self.image_pull_policy,
                     self.request_memory, self.request_cpu, self.limit_memory,
-                    self.limit_cpu, self.limit_gpu, self.gcp_service_account_key, self.node_selectors,
+                    self.limit_cpu, self.gcp_service_account_key, self.node_selectors,
                     self.affinity, self.annotations, self.volumes, self.volume_mounts,
-                    self.tolerations, self.labels)
+                    self.tolerations)
 
     @staticmethod
     def from_dict(obj):
@@ -91,7 +89,6 @@ class KubernetesExecutorConfig:
             request_cpu=namespaced.get('request_cpu', None),
             limit_memory=namespaced.get('limit_memory', None),
             limit_cpu=namespaced.get('limit_cpu', None),
-            limit_gpu=namespaced.get('limit_gpu', None),
             gcp_service_account_key=namespaced.get('gcp_service_account_key', None),
             node_selectors=namespaced.get('node_selectors', None),
             affinity=namespaced.get('affinity', None),
@@ -99,7 +96,6 @@ class KubernetesExecutorConfig:
             volumes=namespaced.get('volumes', []),
             volume_mounts=namespaced.get('volume_mounts', []),
             tolerations=namespaced.get('tolerations', None),
-            labels=namespaced.get('labels', {}),
         )
 
     def as_dict(self):
@@ -110,7 +106,6 @@ class KubernetesExecutorConfig:
             'request_cpu': self.request_cpu,
             'limit_memory': self.limit_memory,
             'limit_cpu': self.limit_cpu,
-            'limit_gpu': self.limit_gpu,
             'gcp_service_account_key': self.gcp_service_account_key,
             'node_selectors': self.node_selectors,
             'affinity': self.affinity,
@@ -118,7 +113,6 @@ class KubernetesExecutorConfig:
             'volumes': self.volumes,
             'volume_mounts': self.volume_mounts,
             'tolerations': self.tolerations,
-            'labels': self.labels,
         }
 
 
@@ -153,7 +147,6 @@ class KubeConfig:
         self.kube_worker_resources = configuration_dict.get('kubernetes_worker_resource_defaults', {})
         self.kube_node_selectors = configuration_dict.get('kubernetes_node_selectors', {})
         self.kube_annotations = configuration_dict.get('kubernetes_annotations', {})
-        self.kube_labels = configuration_dict.get('kubernetes_labels', {})
         self.delete_worker_pods = conf.getboolean(
             self.kubernetes_section, 'delete_worker_pods')
         self.worker_pods_creation_batch_size = conf.getint(
@@ -167,8 +160,8 @@ class KubeConfig:
         self.dags_in_image = conf.getboolean(self.kubernetes_section, 'dags_in_image')
 
         # Run as user for pod security context
-        self.worker_run_as_user = self._get_security_context_val('run_as_user')
-        self.worker_fs_group = self._get_security_context_val('fs_group')
+        self.worker_run_as_user = conf.get(self.kubernetes_section, 'run_as_user')
+        self.worker_fs_group = conf.get(self.kubernetes_section, 'fs_group')
 
         self.dags_configmap = conf.get(self.kubernetes_section, 'dags_configmap')
 
@@ -266,25 +259,7 @@ class KubeConfig:
         else:
             self.kube_tolerations = None
 
-        kube_client_request_args = conf.get(self.kubernetes_section, 'kube_client_request_args')
-        if kube_client_request_args:
-            self.kube_client_request_args = json.loads(kube_client_request_args)
-            if self.kube_client_request_args['_request_timeout'] and \
-                    isinstance(self.kube_client_request_args['_request_timeout'], list):
-                self.kube_client_request_args['_request_timeout'] = \
-                    tuple(self.kube_client_request_args['_request_timeout'])
-        else:
-            self.kube_client_request_args = {}
         self._validate()
-
-    # pod security context items should return integers
-    # and only return a blank string if contexts are not set.
-    def _get_security_context_val(self, scontext):
-        val = configuration.get(self.kubernetes_section, scontext)
-        if len(val) == 0:
-            return val
-        else:
-            return int(val)
 
     def _validate(self):
         # TODO: use XOR for dags_volume_claim and git_dags_folder_mount_point
@@ -312,20 +287,19 @@ class KubeConfig:
 
 
 class KubernetesJobWatcher(multiprocessing.Process, LoggingMixin, object):
-    def __init__(self, namespace, watcher_queue, resource_version, worker_uuid, kube_config):
+    def __init__(self, namespace, watcher_queue, resource_version, worker_uuid):
         multiprocessing.Process.__init__(self)
         self.namespace = namespace
         self.worker_uuid = worker_uuid
         self.watcher_queue = watcher_queue
         self.resource_version = resource_version
-        self.kube_config = kube_config
 
     def run(self):
         kube_client = get_kube_client()
         while True:
             try:
                 self.resource_version = self._run(kube_client, self.resource_version,
-                                                  self.worker_uuid, self.kube_config)
+                                                  self.worker_uuid)
             except Exception:
                 self.log.exception('Unknown error in KubernetesJobWatcher. Failing')
                 raise
@@ -333,7 +307,7 @@ class KubernetesJobWatcher(multiprocessing.Process, LoggingMixin, object):
                 self.log.warn('Watch died gracefully, starting back up with: '
                               'last resource_version: %s', self.resource_version)
 
-    def _run(self, kube_client, resource_version, worker_uuid, kube_config):
+    def _run(self, kube_client, resource_version, worker_uuid):
         self.log.info(
             'Event: and now my watch begins starting at resource_version: %s',
             resource_version
@@ -343,9 +317,6 @@ class KubernetesJobWatcher(multiprocessing.Process, LoggingMixin, object):
         kwargs = {'label_selector': 'airflow-worker={}'.format(worker_uuid)}
         if resource_version:
             kwargs['resource_version'] = resource_version
-        if kube_config.kube_client_request_args:
-            for key, value in kube_config.kube_client_request_args.iteritems():
-                kwargs[key] = value
 
         last_resource_version = None
         for event in watcher.stream(kube_client.list_namespaced_pod, self.namespace,
@@ -412,15 +383,14 @@ class AirflowKubernetesScheduler(LoggingMixin):
         self.kube_client = kube_client
         self.launcher = PodLauncher(kube_client=self.kube_client)
         self.worker_configuration = WorkerConfiguration(kube_config=self.kube_config)
-        self._manager = multiprocessing.Manager()
-        self.watcher_queue = self._manager.Queue()
+        self.watcher_queue = multiprocessing.Queue()
         self.worker_uuid = worker_uuid
         self.kube_watcher = self._make_kube_watcher()
 
     def _make_kube_watcher(self):
         resource_version = KubeResourceVersion.get_current_resource_version()
         watcher = KubernetesJobWatcher(self.namespace, self.watcher_queue,
-                                       resource_version, self.worker_uuid, self.kube_config)
+                                       resource_version, self.worker_uuid)
         watcher.start()
         return watcher
 
@@ -456,18 +426,18 @@ class AirflowKubernetesScheduler(LoggingMixin):
         )
         self.log.debug("Kubernetes pod created: %s", pod)
         # the watcher will monitor pods, so we do not block.
-        self.launcher.run_pod_async(pod, **self.kube_config.kube_client_request_args)
+        self.launcher.run_pod_async(pod)
         self.log.debug("Kubernetes Job created!")
 
     def delete_pod(self, pod_id):
-        try:
-            self.kube_client.delete_namespaced_pod(
-                pod_id, self.namespace, body=client.V1DeleteOptions(),
-                **self.kube_config.kube_client_request_args)
-        except ApiException as e:
-            # If the pod is already deleted
-            if e.status != 404:
-                raise
+        if self.kube_config.delete_worker_pods:
+            try:
+                self.kube_client.delete_namespaced_pod(
+                    pod_id, self.namespace, body=client.V1DeleteOptions())
+            except ApiException as e:
+                # If the pod is already deleted
+                if e.status != 404:
+                    raise
 
     def sync(self):
         """
@@ -479,18 +449,11 @@ class AirflowKubernetesScheduler(LoggingMixin):
 
         """
         self._health_check_kube_watcher()
-        while True:
-            try:
-                task = self.watcher_queue.get_nowait()
-                try:
-                    self.process_watcher_task(task)
-                finally:
-                    self.watcher_queue.task_done()
-            except Empty:
-                break
+        while not self.watcher_queue.empty():
+            self.process_watcher_task()
 
-    def process_watcher_task(self, task):
-        pod_id, state, labels, resource_version = task
+    def process_watcher_task(self):
+        pod_id, state, labels, resource_version = self.watcher_queue.get()
         self.log.info(
             'Attempting to finish pod; pod_id: %s; state: %s; labels: %s',
             pod_id, state, labels
@@ -517,10 +480,10 @@ class AirflowKubernetesScheduler(LoggingMixin):
 
     @staticmethod
     def _make_safe_pod_id(safe_dag_id, safe_task_id, safe_uuid):
-        """
+        r"""
         Kubernetes pod names must be <= 253 chars and must pass the following regex for
         validation
-        "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
+        "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
 
         :param safe_dag_id: a dag_id with only alphanumeric characters
         :param safe_task_id: a task_id with only alphanumeric characters
@@ -639,10 +602,6 @@ class AirflowKubernetesScheduler(LoggingMixin):
         )
         return None
 
-    def terminate(self):
-        self.watcher_queue.join()
-        self._manager.shutdown()
-
 
 class KubernetesExecutor(BaseExecutor, LoggingMixin):
     def __init__(self):
@@ -652,7 +611,6 @@ class KubernetesExecutor(BaseExecutor, LoggingMixin):
         self.kube_scheduler = None
         self.kube_client = None
         self.worker_uuid = None
-        self._manager = multiprocessing.Manager()
         super(KubernetesExecutor, self).__init__(parallelism=self.kube_config.parallelism)
 
     @provide_session
@@ -692,9 +650,6 @@ class KubernetesExecutor(BaseExecutor, LoggingMixin):
                 )
             )
             kwargs = dict(label_selector=dict_string)
-            if self.kube_config.kube_client_request_args:
-                for key, value in self.kube_config.kube_client_request_args.iteritems():
-                    kwargs[key] = value
             pod_list = self.kube_client.list_namespaced_pod(
                 self.kube_config.kube_namespace, **kwargs)
             if len(pod_list.items) == 0:
@@ -715,8 +670,7 @@ class KubernetesExecutor(BaseExecutor, LoggingMixin):
                     self.kube_config.executor_namespace, kubernetes.client.V1Secret(
                         data={
                             'key.json': base64.b64encode(open(secret_path, 'r').read())},
-                        metadata=kubernetes.client.V1ObjectMeta(name=secret_name)),
-                    **self.kube_config.kube_client_request_args)
+                        metadata=kubernetes.client.V1ObjectMeta(name=secret_name)))
             except ApiException as e:
                 if e.status == 409:
                     return self.kube_client.replace_namespaced_secret(
@@ -724,8 +678,7 @@ class KubernetesExecutor(BaseExecutor, LoggingMixin):
                         kubernetes.client.V1Secret(
                             data={'key.json': base64.b64encode(
                                 open(secret_path, 'r').read())},
-                            metadata=kubernetes.client.V1ObjectMeta(name=secret_name)),
-                        **self.kube_config.kube_client_request_args)
+                            metadata=kubernetes.client.V1ObjectMeta(name=secret_name)))
                 self.log.exception(
                     'Exception while trying to inject secret. '
                     'Secret name: %s, error details: %s',
@@ -753,8 +706,8 @@ class KubernetesExecutor(BaseExecutor, LoggingMixin):
         # https://github.com/kubernetes-client/python/blob/master/kubernetes/docs
         # /CoreV1Api.md#list_namespaced_pod
         KubeResourceVersion.reset_resource_version()
-        self.task_queue = self._manager.Queue()
-        self.result_queue = self._manager.Queue()
+        self.task_queue = Queue()
+        self.result_queue = Queue()
         self.kube_client = get_kube_client()
         self.kube_scheduler = AirflowKubernetesScheduler(
             self.kube_config, self.task_queue, self.result_queue,
@@ -779,56 +732,51 @@ class KubernetesExecutor(BaseExecutor, LoggingMixin):
         self.kube_scheduler.sync()
 
         last_resource_version = None
-        while True:
+        while not self.result_queue.empty():
+            results = self.result_queue.get()
+            key, state, pod_id, resource_version = results
+            last_resource_version = resource_version
+            self.log.info('Changing state of %s to %s', results, state)
             try:
-                results = self.result_queue.get_nowait()
-                try:
-                    key, state, pod_id, resource_version = results
-                    last_resource_version = resource_version
-                    self.log.info('Changing state of %s to %s', results, state)
-                    try:
-                        self._change_state(key, state, pod_id)
-                    except Exception as e:
-                        self.log.exception('Exception: %s when attempting ' +
-                                           'to change state of %s to %s, re-queueing.', e, results, state)
-                        self.result_queue.put(results)
-                finally:
-                    self.result_queue.task_done()
-            except Empty:
-                break
+                self._change_state(key, state, pod_id)
+            except Exception as e:
+                self.log.exception('Exception: %s when attempting ' +
+                                   'to change state of %s to %s, re-queueing.', e, results, state)
+                self.result_queue.put(results)
 
         KubeResourceVersion.checkpoint_resource_version(last_resource_version)
 
-        for _ in range(self.kube_config.worker_pods_creation_batch_size):
+        for i in range(min((self.kube_config.worker_pods_creation_batch_size, self.task_queue.qsize()))):
+            task = self.task_queue.get()
+
             try:
-                task = self.task_queue.get_nowait()
-                try:
-                    self.kube_scheduler.run_next(task)
-                except ApiException as e:
-                    self.log.warning('ApiException when attempting to run task, re-queueing. '
-                                     'Message: %s' % json.loads(e.body)['message'])
-                    self.task_queue.put(task)
-                finally:
-                    self.task_queue.task_done()
-            except Empty:
-                break
+                self.kube_scheduler.run_next(task)
+            except ApiException:
+                self.log.exception('ApiException when attempting ' +
+                                   'to run task, re-queueing.')
+                self.task_queue.put(task)
 
     def _change_state(self, key, state, pod_id):
         if state != State.RUNNING:
-            if self.kube_config.delete_worker_pods:
-                self.kube_scheduler.delete_pod(pod_id)
-                self.log.info('Deleted pod: %s', str(key))
+            self.kube_scheduler.delete_pod(pod_id)
             try:
+                self.log.info('Deleted pod: %s', str(key))
                 self.running.pop(key)
             except KeyError:
                 self.log.debug('Could not find key: %s', str(key))
                 pass
         self.event_buffer[key] = state
+        (dag_id, task_id, ex_time, try_number) = key
+        with create_session() as session:
+            item = session.query(TaskInstance).filter_by(
+                dag_id=dag_id,
+                task_id=task_id,
+                execution_date=ex_time
+            ).one()
+            if state:
+                item.state = state
+                session.add(item)
 
     def end(self):
         self.log.info('Shutting down Kubernetes executor')
         self.task_queue.join()
-        self.result_queue.join()
-        if self.kube_scheduler:
-            self.kube_scheduler.terminate()
-        self._manager.shutdown()
